src/requirements.txt
discord
openai==1.3.0
python-dotenv
requests


